Factor out assertValid method
